requirements.txt
environs
requests
pillow
selenium
loguru